requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
phonenumbers==8.13.50
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.0
pydantic-extra-types==2.10.0
pydantic_core==2.27.0
Pygments==2.18.0
PyJWT==2.10.0
PyMySQL==1.1.1
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
typer==0.13.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==0.24.0
websockets==14.1
